8.
////////////////////////////////////////////////////////
// C++

1) InventoryComponent
- 일반 아이템 사용, 장비 아이템 착용, 아이템 퀵 슬롯에 등록, 등록된 퀵 아이템 사용, 아이템 이동, 아이템 스왑, 아이템 정렬, 아이템 잠금, 아이템 파괴 기능이 추가됨.

2) PlayerComponentInterface
- BP에서 ActorCompoenent의 함수들을 손쉽게 접근하기 위한 클래스
- SoulCharacter에서 구현되어있다.
- BP에서 GetControlledPawn이나 Pawn에서 사용가능함

3) SoulCharacter
- PlayerComponentInterface의 함수들을 구현함
- Add, Remove, Swap, Move, Sort, Use, Equip, UnEquip 등등

4) SoulPlayerController
- WB_Equipment와 WB_Inventory의 슬롯들을 갱신, 초기화

////////////////////////////////////////////////////////
// HUD

1) Inventory
- 인벤토리 내 아이템 사용, 퀵 등록, 파괴, 잠금이 가능해짐
- 아이템 ID에 따른 정렬이 가능함
- DragDrop으로 아이템 슬롯들간의 위치이동이 가능함.

2) Equipment
- 장착 현황 및 스테이터스를 보여줌
- 아이템 장착/장착 해제에 따른 변화

+++ Source/Soulslike/Public/Player/StatComponent.h
@@ -64,7 +64,7 @@ public:
 	FORCEINLINE float GetMaxHp() const { return PlayerStat.MaxHp; }
 	FORCEINLINE float GetCurStamina() const { return PlayerStat.CurStamina; }
 	FORCEINLINE float GetMaxStamina() const { return PlayerStat.MaxStamina; }
-	FORCEINLINE float GetSoulsCount() const { return PlayerStat.SoulsCount; }
+	FORCEINLINE int32 GetSoulsCount() const { return PlayerStat.SoulsCount; }
 
 	virtual void GetLifetimeReplicatedProps(TArray<FLifetimeProperty>& OutLifetimeProps) const override;
 };
